requirements.txt
streamlit
opencv-python-headless
numpy
pandas
scikit-image
scikit-learn
pillow
